# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SupportApi",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "C:\\Users\\a\\OneDrive\\Desktop\\support\\SupportApi\\app"
        }
    ],
    "paths": {
...
        "/api/v2/license-request/store": {
            "post": {
                "tags": [
                    "License"
                ],
                "summary": "Store a license request",
                "operationId": "9b84524f535e4fde483f096fc41ccf4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/license-request": {
            "get": {
                "tags": [
                    "License"
                ],
                "summary": "Get all licenses",
                "operationId": "c3cec27d98ba8b63d4377913a55fe6ce",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Failed response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/license-request/LicenseRequestByClient/{id}": {
            "get": {
                "tags": [
                    "License"
                ],
                "summary": "LicenseRequestByClient",
                "operationId": "c09fae4f9ca4fdea8743d34d5aa5f297",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Failed response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/license-request/LicenseRequestByAssignedPerson/{id}": {
            "get": {
                "tags": [
                    "License"
                ],
                "summary": "LicenseRequestByAssignedPerson",
                "operationId": "30ae517c53d0ca118739efa766c270fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the assigned person",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Failed response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/license-request/send": {
            "post": {
                "tags": [
                    "License"
                ],
                "summary": "Update license request",
                "operationId": "updateLicenseRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Failed response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/license-request/approve": {
            "post": {
                "tags": [
                    "License"
                ],
                "summary": "Approve license request",
                "operationId": "approveLicenseRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Failed response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/license-request/disapprove": {
            "post": {
                "tags": [
                    "License"
                ],
                "summary": "Disapprove license request",
                "operationId": "disapproveLicenseRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Failed response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}